requirements.txt
Django
mysql-python
django_digg_paginator
lastfmclient
-e git://github.com/alastair/python-musicbrainzngs.git#egg=musicbrainzngs